updates
created the custom linked list so that we can implement a linear search
on the linked list, fixed an issue where the program did not compiled
and fixed the references between the abstract class of person and user.
program is compiling!!!

missing:

testing on the user thing.
also finish adding documentation.
search on the linked list.